Removed old IAP view
