+++ lib/app/router.gr.dart
@@ -8,7 +8,6 @@
 
 import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
-import 'package:lta_datamall_flutter/ui/views/iap/purchase_view1.dart';
 
 import '../ui/views/bus/bus_arrival/bus_arrival_view.dart';
 import '../ui/views/bus/bus_view.dart';
@@ -17,13 +16,10 @@ import '../ui/views/iap/purchase_view.dart';
 class Routes {
   static const String busView = '/';
   static const String busArrivalView = '/bus-arrival-view';
-  static const String marketScreen = '/market-screen';
   static const String purchaseView = '/purchase-view';
-
   static const all = <String>{
     busView,
     busArrivalView,
-    marketScreen,
     purchaseView,
   };
 }
@@ -34,7 +30,6 @@ class Router extends RouterBase {
   final _routes = <RouteDef>[
     RouteDef(Routes.busView, page: BusView),
     RouteDef(Routes.busArrivalView, page: BusArrivalView),
-    RouteDef(Routes.marketScreen, page: MarketScreen),
     RouteDef(Routes.purchaseView, page: PurchaseView),
   ];
   @override
@@ -53,12 +48,6 @@ class Router extends RouterBase {
           busStopCode: args.busStopCode,
           description: args.description,
         ),
-        settings: data,
-      );
-    },
-    MarketScreen: (data) {
-      return MaterialPageRoute<dynamic>(
-        builder: (context) => MarketScreen(),
         settings: data,
       );
     },
